Panaversity_Course/OpenAI_AgentSDK/Q4_AgenticAI_Repo/01_ai_agents_first/10_streaming/hello_streaming/pyproject.toml
[project]
name = "hello-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
